+++ Quicksilver/Code-App/QSMainPreferencePanes.h
@@ -38,7 +38,7 @@
 @interface QSActionsPrefPane : QSPreferencePane {
 	NSMutableArray *actions;
 	NSMutableArray *groups;
-	int displayMode;
+	NSInteger displayMode;
 	IBOutlet NSArrayController *groupController;
 	IBOutlet NSArrayController *actionController;
 }
@@ -46,8 +46,8 @@
 - (void)setActions:(NSMutableArray *)newActions;
 - (NSMutableArray *)groups;
 - (void)setGroups:(NSMutableArray *)newGroups;
-- (int) displayMode;
-- (void)setDisplayMode:(int)newDisplayMode;
+- (NSInteger) displayMode;
+- (void)setDisplayMode:(NSInteger)newDisplayMode;
 - (IBAction)setFilterText:(id)sender;
 - (void)updateGroups;
 - (void)selectCategories:(NSArray *)categories;
