+++ Quicksilver/Code-QuickStepEffects/QSCGSTransition.h
@@ -14,9 +14,9 @@
 	CGSTransitionSpec spec;
 }
 + (id)transitionWithWindow:(NSWindow *)window type:(CGSTransitionType)type option:(CGSTransitionOption)option;
-+ (id)transitionWithWindow:(NSWindow *)window type:(CGSTransitionType)type option:(CGSTransitionOption)option duration:(float)duration;
-+ (id)transitionWithType:(CGSTransitionType)type option:(CGSTransitionOption)option duration:(float)duration;
++ (id)transitionWithWindow:(NSWindow *)window type:(CGSTransitionType)type option:(CGSTransitionOption)option duration:(CGFloat)duration;
++ (id)transitionWithType:(CGSTransitionType)type option:(CGSTransitionOption)option duration:(CGFloat)duration;
 - (id)initWithType:(CGSTransitionType) type option:(CGSTransitionOption)option;
 - (void)attachToWindow:(NSWindow *)window;
-- (void)runTransition:(float)duration;
+- (void)runTransition:(CGFloat)duration;
 @end
