+++ Quicksilver/Code-QuickStepFoundation/NSColor_QSModifications.m
@@ -10,13 +10,13 @@
 
 @implementation NSColor (Contrast)
 
-- (NSColor *)colorWithLighting:(float)light { return [self colorWithLighting:light plasticity:0]; }
+- (NSColor *)colorWithLighting:(CGFloat)light { return [self colorWithLighting:light plasticity:0]; }
 
-- (NSColor *)colorWithLighting:(float)light plasticity:(float)plastic {
+- (NSColor *)colorWithLighting:(CGFloat)light plasticity:(CGFloat)plastic {
 	if (plastic > 1) plastic = 1.0;
 	if (plastic < 0) plastic = 0.0;
 	NSColor *color = [self colorUsingColorSpaceName:NSCalibratedRGBColorSpace];
-	float h, s, b, a;
+	CGFloat h, s, b, a;
 
 	[color getHue:&h saturation:&s brightness:&b alpha:&a];
